moved components into one file, added Navbar and Footer to MainPage
