--- tech view/src/App.js
@@ -1,22 +1,0 @@
-import Navbar from "./Navbar";
-import Footer from "./Footer";
-import React from 'react';
-// import { Route, Routes } from "react-router-dom";
-
-function App() {
-  return (
-    <div>
-      <Navbar />
-      <div>
-        <Footer />
-      </div>
-    </div>
-
-
-  )
-}
-
-
-
-export default App;
-
