--- tech view/src/components/Pricing.js
@@ -1,3 +1,0 @@
-export default function Pricing() {
-  return <h1>Pricing</h1>
-}
