+++ tech-view/src/App.js
@@ -1,4 +1,3 @@
-import './App.css'
 import MainPage from './components/MainPage'
 
 function App () {
