+++ tech-view/src/components/Community.js
@@ -1,0 +1,3 @@
+export default function About () {
+  return <h1>About</h1>
+}
