+++ tech-view/src/components/Footer.css
