+++ tech-view/src/components/Home.js
@@ -1,0 +1,3 @@
+export default function Home () {
+  return <h1>Home</h1>
+}
