+++ tech-view/src/components/MainPage.js
@@ -1,11 +1,15 @@
+import Navbar from './Navbar'
 import PostCardsContainer from './PostCardsContainer'
 import SideBar from './SideBar'
+import Footer from './Footer'
 
 export default function MainPage () {
   return (
     <>
+      <Navbar />
       <SideBar />
       <PostCardsContainer />
+      <Footer />
     </>
   )
 }
