+++ tech-view/src/components/Navbar.css
