+++ tech-view/src/components/Pricing.js
@@ -1,0 +1,3 @@
+export default function Pricing () {
+  return <h1>Pricing</h1>
+}
