+++ tech-view/src/components/logo.png
